@@ -45,36 +44,29 @@
         "vue-i18n": "9.3.0-beta.19"
     },
     "devDependencies": {
-        "@aerogel/cli": "0.0.0-next.13b88f634a31caf3669d61e53e7d3fa95e20488c",
-        "@aerogel/cypress": "0.0.0-next.ecf4c2737b43d20cfac283ce80ac5913de65f502",
-        "@aerogel/histoire": "0.0.0-next.13b88f634a31caf3669d61e53e7d3fa95e20488c",
-        "@aerogel/vite": "0.0.0-next.ad39a2a31f43d07c0518ac80d8b1a8e2f823027d",
+        "@aerogel/cli": "0.0.0-next.44f5953d2450ac149ceabae6448a967b9def9d2a",
+        "@aerogel/cypress": "0.0.0-next.88677f04a9d8fa694eab48c30f9a56f59df19206",
+        "@aerogel/vite": "0.0.0-next.24ed4c26fdb36195386fee4a81543fb84f51c979",
         "@iconify/json": "^2.2.134",
-        "@noeldemartin/eslint-config-vue": "next",
-        "@noeldemartin/scripts": "next",
+        "@noeldemartin/eslint-config-vue": "0.1.1-next.8da46d7532acbe6ee4126f410141b7f281ec349a",
+        "@noeldemartin/scripts": "0.3.0-next.2dfd366e59e45ecf5ead1a617e4d0e870dbea507",
         "@solid/community-server": "^7.1.6",
         "@total-typescript/ts-reset": "^0.4.2",
+        "@tsconfig/node22": "^22.0.1",
         "@types/node": "^20.3.1",
+        "@vue/tsconfig": "^0.7.0",
         "autoprefixer": "^10.4.14",
         "concurrently": "^8.2.0",
-        "cypress": "^12.17.0",
-        "eslint": "^8.40.0",
-        "histoire": "0.17.6",
-        "patch-package": "^8.0.0",
+        "cypress": "^14.2.0",
         "postcss-pseudo-classes": "^0.4.0",
-        "prettier": "^2.8.8",
-        "prettier-eslint-cli": "^7.1.0",
         "prettier-plugin-tailwindcss": "^0.2.8",
         "start-server-and-test": "^2.0.0",
-        "typescript": "~5.4",
+        "typescript": "^5.8.2",
         "unplugin-icons": "^0.16.3",
         "unplugin-vue-components": "^0.24.1",
-        "vite": "^4.3.0",
-        "vitest": "^0.33.0",
-        "vue-tsc": "^1.8.15"
-    },
-    "overrides": {
-        "@noeldemartin/solid-utils": "0.5.0-next.612e9e93903ec33bbcba7105ed2a8580ed2d754c"
+        "vite": "^6.2.2",
+        "vitest": "^3.0.9",
+        "vue-tsc": "^2.2.8"
     },
     "eslintConfig": {
         "extends": [
