@@ -12,4 +12,4 @@ if (process.env.NODE_ENV === 'development') {
     };
 }
 
-module.exports = { plugins };
+export default { plugins };
