@@ -1,5 +1,5 @@
 import i18n from '@aerogel/plugin-i18n';
-import offlineFirst from '@aerogel/plugin-offline-first';
+import localFirst from '@aerogel/plugin-local-first';
 import routing from '@aerogel/plugin-routing';
 import solid from '@aerogel/plugin-solid';
 import soukai from '@aerogel/plugin-soukai';
@@ -20,7 +20,7 @@ bootstrap(App, {
         soukai({ models: import.meta.glob(['@/models/*', '!**/*.test.ts'], { eager: true }) }),
         routing({ routes, bindings }),
         solid(),
-        offlineFirst(),
+        localFirst(),
     ],
     install(app) {
         app.use(vivant());
