@@ -57,7 +57,7 @@
 </template>
 
 <script setup lang="ts">
-import { Cloud } from '@aerogel/plugin-offline-first';
+import { Cloud } from '@aerogel/plugin-local-first';
 import { computed } from 'vue';
 import { UI, translate } from '@aerogel/core';
 
