@@ -42,7 +42,7 @@
 </template>
 
 <script setup lang="ts">
-import { Cloud } from '@aerogel/plugin-offline-first';
+import { Cloud } from '@aerogel/plugin-local-first';
 import { computed } from 'vue';
 import { required, urlParse } from '@noeldemartin/utils';
 import { Solid } from '@aerogel/plugin-solid';
